added tests for select menu
