+++ src/__tests__/unit/components/form-components/select/select.test.tsx
@@ -497,300 +497,6 @@ describe('Select', () => {
     expect(focusedOption).toHaveFocus();
   });
 
-  it(`sets the value of the field when the user clicks on an option in the 
-  menu.`, async () => {
-    const favoriteColor = new Field({
-      name: 'favoriteColor',
-      defaultValue: '',
-    });
-
-    const options = [
-      {
-        text: 'Red',
-        value: 'red',
-      },
-      {
-        text: 'Green',
-        value: 'green',
-      },
-      {
-        text: 'Blue',
-        value: 'blue',
-      },
-    ];
-
-    render(
-      <Select label="Favorite Color" field={favoriteColor} options={options} />,
-    );
-
-    const combobox = screen.getByRole('combobox');
-    await user.click(combobox);
-
-    const optionElements = screen.getAllByRole('option');
-    await user.click(optionElements[2]);
-
-    expect(favoriteColor.state.value).toBe(options[2].value);
-  });
-
-  it(`sets the value of the field and restores focus to the combobox when the 
-  user presses Enter on an option in the menu.`, async () => {
-    const favoriteColor = new Field({
-      name: 'favoriteColor',
-      defaultValue: '',
-    });
-
-    const options = [
-      {
-        text: 'Red',
-        value: 'red',
-      },
-      {
-        text: 'Green',
-        value: 'green',
-      },
-      {
-        text: 'Blue',
-        value: 'blue',
-      },
-    ];
-
-    render(
-      <Select label="Favorite Color" field={favoriteColor} options={options} />,
-    );
-
-    const combobox = screen.getByRole('combobox');
-    combobox.focus();
-    await user.keyboard('g');
-    await user.keyboard('{Enter}');
-
-    expect(favoriteColor.state.value).toBe(options[1].value);
-    expect(combobox).toHaveFocus();
-  });
-
-  it(`sets the value of the field and restores focus to the combobox when the 
-  user presses Tab on an option in the menu.`, async () => {
-    const favoriteColor = new Field({
-      name: 'favoriteColor',
-      defaultValue: '',
-    });
-
-    const options = [
-      {
-        text: 'Red',
-        value: 'red',
-      },
-      {
-        text: 'Green',
-        value: 'green',
-      },
-      {
-        text: 'Blue',
-        value: 'blue',
-      },
-    ];
-
-    render(
-      <Select label="Favorite Color" field={favoriteColor} options={options} />,
-    );
-
-    const combobox = screen.getByRole('combobox');
-    combobox.focus();
-    await user.keyboard('g');
-    await user.keyboard('{Tab}');
-
-    expect(favoriteColor.state.value).toBe(options[1].value);
-    expect(combobox).toHaveFocus();
-  });
-
-  it(`closes the menu and restores focus to the combobox without changing the 
-  value of the field when the user presses the Escape key while an option is 
-  selected.`, async () => {
-    const favoriteColor = new Field({
-      name: 'favoriteColor',
-      defaultValue: '',
-    });
-
-    render(
-      <Select
-        label="Favorite Color"
-        field={new Field({ name: 'favoriteColor', defaultValue: '' })}
-        options={[
-          {
-            text: 'Red',
-            value: 'red',
-          },
-          {
-            text: 'Green',
-            value: 'green',
-          },
-          {
-            text: 'Blue',
-            value: 'blue',
-          },
-        ]}
-      />,
-    );
-
-    const menuContainer = document.getElementsByClassName('menu_container')[0];
-    expect(menuContainer.classList).toContain('hidden');
-
-    const combobox = screen.getByRole('combobox');
-    await user.click(combobox);
-
-    const menuOptions = screen.getAllByRole('option');
-    expect(menuOptions[0]).toHaveFocus();
-
-    await user.keyboard('{Escape}');
-    expect(favoriteColor.state.value).toBe('');
-  });
-
-  it(`navigates to the next option in the menu when the user presses the 
-  ArrowDown key while an option is in focus, stopping on the last option if 
-  the user continues to press the ArrowDown key.`, async () => {
-    const options = [
-      {
-        text: 'Red',
-        value: 'red',
-      },
-      {
-        text: 'Green',
-        value: 'green',
-      },
-      {
-        text: 'Blue',
-        value: 'blue',
-      },
-      {
-        text: 'Yellow',
-        value: 'yellow',
-      },
-      {
-        text: 'Purple',
-        value: 'purple',
-      },
-      {
-        text: 'Orange',
-        value: 'orange',
-      },
-    ];
-
-    render(
-      <Select
-        label="Favorite Color"
-        field={new Field({ name: 'favoriteColor', defaultValue: '' })}
-        options={options}
-      />,
-    );
-
-    const combobox = screen.getByRole('combobox');
-    const menuOptions = screen.getAllByRole('option');
-    combobox.focus();
-
-    for (let i = 0; i <= options.length + 10; i++) {
-      await user.keyboard('{ArrowDown}');
-      const focusedOptionIndex = Math.min(i, options.length - 1);
-      expect(menuOptions[focusedOptionIndex]).toHaveFocus();
-    }
-  });
-
-  it(`navigates to the previous option in the menu when the user presses the 
-  ArrowUp key while an option is in focus, stopping on the first option if 
-  the user continues to press the ArrowUp key.`, async () => {
-    const options = [
-      {
-        text: 'Red',
-        value: 'red',
-      },
-      {
-        text: 'Green',
-        value: 'green',
-      },
-      {
-        text: 'Blue',
-        value: 'blue',
-      },
-      {
-        text: 'Yellow',
-        value: 'yellow',
-      },
-      {
-        text: 'Purple',
-        value: 'purple',
-      },
-      {
-        text: 'Orange',
-        value: 'orange',
-      },
-    ];
-
-    render(
-      <Select
-        label="Favorite Color"
-        field={new Field({ name: 'favoriteColor', defaultValue: '' })}
-        options={options}
-      />,
-    );
-
-    const combobox = screen.getByRole('combobox');
-    const menuOptions = screen.getAllByRole('option');
-    combobox.focus();
-
-    for (let i = options.length - 1; i >= -10; i--) {
-      await user.keyboard('{ArrowUp}');
-      const focusedOptionIndex = Math.max(i, 0);
-      expect(menuOptions[focusedOptionIndex]).toHaveFocus();
-    }
-  });
-
-  it(`navigates to the first option beginning with a given character when the 
-  user presses the corresponding key while any option in the menu has 
-  focus.`, async () => {
-    const options = [
-      {
-        text: 'Red',
-        value: 'red',
-      },
-      {
-        text: 'Green',
-        value: 'green',
-      },
-      {
-        text: 'Blue',
-        value: 'blue',
-      },
-      {
-        text: 'Yellow',
-        value: 'yellow',
-      },
-      {
-        text: 'Purple',
-        value: 'purple',
-      },
-      {
-        text: 'Orange',
-        value: 'orange',
-      },
-    ];
-
-    render(
-      <Select
-        label="Favorite Color"
-        field={new Field({ name: 'favoriteColor', defaultValue: '' })}
-        options={options}
-      />,
-    );
-
-    const combobox = screen.getByRole('combobox');
-    const menuOptions = screen.getAllByRole('option');
-    combobox.focus();
-
-    for (let i = 0; i < options.length; i++) {
-      await user.keyboard(options[i].value[0]);
-      const focusedOptionIndex = i;
-      expect(menuOptions[focusedOptionIndex]).toHaveFocus();
-    }
-  });
-
   it(`adds any classNames it receives to the classList of the outer 
   container.`, () => {
     const className = 'test';
@@ -1088,7 +794,7 @@ describe('Select', () => {
     isMenuScrolledToBottomSpy.mockRestore();
   });
 
-  it(`hides the scroll up buttom and shows the scroll down button when the
+  it(`hides the scroll up button and shows the scroll down button when the
   screen is resized, the menu is scrollable, and the menu is scrolled to the
   top.`, async () => {
     let isMenuScrolledToTop = false;
@@ -1262,184 +968,6 @@ describe('Select', () => {
     isMenuScrolledToBottomSpy.mockRestore();
   });
 
-  it(`hides the scroll up buttom and shows the scroll down button when the
-  menu is scrolled and reaches the top.`, async () => {
-    let isMenuScrolledToTop = false;
-    let isMenuScrolledToBottom = true;
-
-    const isMenuScrollableSpy = jest
-      .spyOn(scrollUtils, 'isMenuScrollable')
-      .mockImplementation(() => true);
-
-    const isMenuScrolledToTopSpy = jest
-      .spyOn(scrollUtils, 'isMenuScrolledToTop')
-      .mockImplementation(() => isMenuScrolledToTop);
-
-    const isMenuScrolledToBottomSpy = jest
-      .spyOn(scrollUtils, 'isMenuScrolledToBottom')
-      .mockImplementation(() => isMenuScrolledToBottom);
-
-    render(
-      <Select
-        label="Select an option"
-        field={new Field({ name: 'favoriteColor', defaultValue: '' })}
-        options={[
-          {
-            text: 'Red',
-            value: 'red',
-          },
-          {
-            text: 'Green',
-            value: 'green',
-          },
-          {
-            text: 'Blue',
-            value: 'blue',
-          },
-        ]}
-      />,
-    );
-
-    await user.click(screen.getByRole('combobox'));
-    await waitFor(() =>
-      expect(screen.queryByAltText('Scroll up')).toBeInTheDocument(),
-    );
-    expect(screen.queryByAltText('Scroll down')).not.toBeInTheDocument();
-
-    isMenuScrolledToTop = true;
-    isMenuScrolledToBottom = false;
-
-    const menu = screen.getByRole('listbox');
-    fireEvent.scroll(menu);
-
-    await waitFor(() => {
-      expect(screen.queryByAltText('Scroll up')).not.toBeInTheDocument();
-      expect(screen.queryByAltText('Scroll down')).toBeInTheDocument();
-    });
-
-    isMenuScrollableSpy.mockRestore();
-    isMenuScrolledToTopSpy.mockRestore();
-    isMenuScrolledToBottomSpy.mockRestore();
-  });
-
-  it(`hides the scroll down button and shows the scroll up button when the
-  menu is scrolled and reaches the bottom.`, async () => {
-    let isMenuScrolledToTop = true;
-    let isMenuScrolledToBottom = false;
-
-    const isMenuScrollableSpy = jest
-      .spyOn(scrollUtils, 'isMenuScrollable')
-      .mockImplementation(() => true);
-
-    const isMenuScrolledToTopSpy = jest
-      .spyOn(scrollUtils, 'isMenuScrolledToTop')
-      .mockImplementation(() => isMenuScrolledToTop);
-
-    const isMenuScrolledToBottomSpy = jest
-      .spyOn(scrollUtils, 'isMenuScrolledToBottom')
-      .mockImplementation(() => isMenuScrolledToBottom);
-
-    render(
-      <Select
-        label="Select an option"
-        field={new Field({ name: 'favoriteColor', defaultValue: '' })}
-        options={[
-          {
-            text: 'Red',
-            value: 'red',
-          },
-          {
-            text: 'Green',
-            value: 'green',
-          },
-          {
-            text: 'Blue',
-            value: 'blue',
-          },
-        ]}
-      />,
-    );
-
-    await user.click(screen.getByRole('combobox'));
-    await waitFor(() =>
-      expect(screen.queryByAltText('Scroll down')).toBeInTheDocument(),
-    );
-    expect(screen.queryByAltText('Scroll up')).not.toBeInTheDocument();
-
-    isMenuScrolledToTop = false;
-    isMenuScrolledToBottom = true;
-
-    const menu = screen.getByRole('listbox');
-    fireEvent.scroll(menu);
-
-    await waitFor(() => {
-      expect(screen.queryByAltText('Scroll down')).not.toBeInTheDocument();
-      expect(screen.queryByAltText('Scroll up')).toBeInTheDocument();
-    });
-
-    isMenuScrollableSpy.mockRestore();
-    isMenuScrolledToTopSpy.mockRestore();
-    isMenuScrolledToBottomSpy.mockRestore();
-  });
-
-  it(`shows both scroll buttons when the menu is scrolled without reaching the
-  top or bottom.`, async () => {
-    let isMenuScrolledToTop = true;
-
-    const isMenuScrollableSpy = jest
-      .spyOn(scrollUtils, 'isMenuScrollable')
-      .mockImplementation(() => true);
-
-    const isMenuScrolledToTopSpy = jest
-      .spyOn(scrollUtils, 'isMenuScrolledToTop')
-      .mockImplementation(() => isMenuScrolledToTop);
-
-    const isMenuScrolledToBottomSpy = jest
-      .spyOn(scrollUtils, 'isMenuScrolledToBottom')
-      .mockImplementation(() => false);
-
-    render(
-      <Select
-        label="Select an option"
-        field={new Field({ name: 'favoriteColor', defaultValue: '' })}
-        options={[
-          {
-            text: 'Red',
-            value: 'red',
-          },
-          {
-            text: 'Green',
-            value: 'green',
-          },
-          {
-            text: 'Blue',
-            value: 'blue',
-          },
-        ]}
-      />,
-    );
-
-    await user.click(screen.getByRole('combobox'));
-    await waitFor(() =>
-      expect(screen.queryByAltText('Scroll down')).toBeInTheDocument(),
-    );
-    expect(screen.queryByAltText('Scroll up')).not.toBeInTheDocument();
-
-    isMenuScrolledToTop = false;
-
-    const menu = screen.getByRole('listbox');
-    fireEvent.scroll(menu);
-
-    await waitFor(() => {
-      expect(screen.queryByAltText('Scroll up')).toBeInTheDocument();
-    });
-    expect(screen.queryByAltText('Scroll down')).toBeInTheDocument();
-
-    isMenuScrollableSpy.mockRestore();
-    isMenuScrolledToTopSpy.mockRestore();
-    isMenuScrolledToBottomSpy.mockRestore();
-  });
-
   it(`sets MenuControls.isKeyboardNavigating.current to false when the user
   moves the mouse.`, async () => {
     const isKeyboardNavigating = {
@@ -1476,35 +1004,4 @@ describe('Select', () => {
 
     useMenuSpy.mockRestore();
   });
-
-  it(`calls MenuControls.startScrollingUp when the mouse enters the scroll up
-  button and the user is not navigating with the keyboard.`, async () => {});
-
-  it(`does not call MenuControls.startScrollingUp when the mouse enters the
-  scroll up button and the user is navigating with the keyboard.`, () => {});
-
-  // it(`calls MenuControls.stopScrolling when the mouse leaves the scroll up
-  // button.`, () => {});
-
-  // it(`calls MenuControls.startScrollingUp when the mousedown event of the scroll
-  // up button is fired, and then calls stopScrolling when the mouseup event is
-  // fired.`, () => {});
-
-  // it(`calls MenuControls.startScrollingUp when the touchstart event of the
-  // scroll up button is fired, and then calls stopScrolling when the touchend
-  // event is fired.`, () => {});
-
-  // it(`calls MenuControls.startScrollingUp when the touchstart event of the
-  // scroll up button is fired, and then calls stopScrolling when the touchcancel
-  // event is fired.`, () => {});
-
-  // it(`calls MenuControls.startScrollingUp when the touchstart event of the
-  // scroll up button is fired, and then calls stopScrolling when the touchmove
-  // event fires and the touch coordinates have moved outside of the
-  // button.`, () => {});
-
-  // it(`calls MenuControls.startScrollingUp when the touchstart event of the
-  // scroll up button is fired but does not call stopScrolling when the touchmove
-  // event fires and the touch coordinates have remained inside of the
-  // button.`, () => {});
 });
